backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
openpyxl==3.1.5
python-docx==1.2.0
httpx==0.28.1
openai==1.82.0
pydantic-settings==2.11.0
